requirements-dev.txt
pep8 >= 1.5.1
flake8 >= 2.1.0
nose >= 1.3.1
coverage >= 3.7.1
nosexcover >= 1.0.10
python-coveralls >= 2.4.2
doublex >= 1.8.1
pexpect >= 3.2
releaseme
wheel
factory_boy

pytest >= 2.6.1
pytest-cov >= 1.7.0
pytest-xdist >= 1.10
pytest-django >= 2.6.2
